+++ src/main/java/me/flotsam/frettler/instrument/FrettedInstrument.java
@@ -17,19 +17,22 @@
 
 package me.flotsam.frettler.instrument;
 
+import static me.flotsam.frettler.engine.Scale.CHROMATIC_SCALE;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import lombok.Getter;
+import me.flotsam.frettler.engine.IntervalPattern;
 import me.flotsam.frettler.engine.Note;
 import me.flotsam.frettler.engine.Scale;
+import me.flotsam.frettler.engine.ScaleInterval;
 import me.flotsam.frettler.engine.ScaleNote;
 
 public abstract class FrettedInstrument {
-
-  public static final int DEFAULT_FRETS = 12;
+  
+  static final int DEFAULT_FRETS = 12;
 
   @Getter
   List<Fret> allFrets = new ArrayList<>();
@@ -56,36 +59,31 @@ public abstract class FrettedInstrument {
     this.frets = frets;
     this.stringNotes = Arrays.asList(strings);
 
+    Scale referenceScale = new Scale(strings[0], IntervalPattern.SCALE_CHROMATIC);
+    int accumulativeDistance = 0;
+    ScaleNote prevStringNote = referenceScale.getHead();
+
     for (int stringNum = 0; stringNum < stringNotes.size(); stringNum++) {
-      Optional<ScaleNote> optScaleNote =
-          Scale.CHROMATIC_SCALE.findScaleNote(stringNotes.get(stringNum));
-      ScaleNote scaleNote = optScaleNote.get();
-
+      ScaleNote stringNote =
+          referenceScale.findScaleNote(stringNotes.get(stringNum)).get();
+      ScaleNote fretNote = stringNote; 
+      
+      accumulativeDistance += calcRelativeDistance(prevStringNote, stringNote);
       for (int fretNum = 0; fretNum <= frets; fretNum++) {
-
-        int octave = 0; // until known from prev fret
+        
         Fret fret = null;
 
+        int fretDistance = accumulativeDistance + fretNum;
+        int currentOctave = fretDistance / ScaleInterval.P8.getSemiTones();
+
         if (stringNum == 0 && (fretNum > 0 && fretNum < 6) && isBanjo()) {
-          fret = new Fret(stringNum * frets + fretNum, null, octave, stringNum,
+          fret = new Fret(stringNum * frets + fretNum, null, currentOctave, stringNum,
               stringNotes.get(stringNum), fretNum);
         } else {
-          // first look for relative fret note on current string
-          Fret prevFret = findRelativeFretUp(stringNum, scaleNote.getNote());
-          if (prevFret != null) {
-            octave = prevFret.getOctave() + 1;
-          } else {
-            // otherwise on the prev string
-            prevFret = findRelativeFretUp(stringNum - 1, scaleNote.getNote());
-            if (prevFret != null) {
-              octave =
-                  prevFret.getFretNum() > fretNum ? prevFret.getOctave() : prevFret.getOctave() + 1;
-            }
-          }
-          fret = new Fret(stringNum * frets + fretNum, scaleNote.getNote(), octave, stringNum,
+          fret = new Fret(stringNum * frets + fretNum, fretNote.getNote(), currentOctave, stringNum,
               stringNotes.get(stringNum), fretNum);
 
-          scaleNote = scaleNote.getNextScaleNote();
+          fretNote = fretNote.getNextScaleNote();
         }
         allFrets.add(fret);
 
@@ -100,9 +98,21 @@ public abstract class FrettedInstrument {
       final int x = stringNum;
       fretsByString
           .add(allFrets.stream().filter(t -> t.getStringNum() == x).collect(Collectors.toList()));
+
+      prevStringNote = stringNote;
     }
   }
 
+  
+  private int calcRelativeDistance(ScaleNote fromNote, ScaleNote toNote) {
+    // need to calculate distance from the from note in a new scale starting at that note
+    // to avoid the wrapping issue if the to note occurs  before the from note 
+    Scale referenceScale = new Scale(fromNote.getNote(), IntervalPattern.SCALE_CHROMATIC);
+    ScaleNote from = referenceScale.findScaleNote(fromNote.getNote()).get();
+    ScaleNote to = referenceScale.findScaleNote(toNote.getNote()).get();
+    return to.getInterval().get().getSemiTones() - from.getInterval().get().getSemiTones();
+  }
+  
   public boolean isBanjo() {
     return false;
   }
@@ -111,19 +121,4 @@ public abstract class FrettedInstrument {
     return fretsByString.size();
   }
 
-  private Fret findRelativeFretUpFromIndex(int stringNum, Note note, int index) {
-    Fret fret = null;
-    for (int idx = index; idx-- > 0;) {
-      Fret aFret = allFrets.get(idx);
-      if (aFret.getNote() == note && aFret.getStringNum() == stringNum) {
-        fret = allFrets.get(idx);
-        break;
-      }
-    }
-    return fret;
-  }
-
-  private Fret findRelativeFretUp(int stringNum, Note note) {
-    return findRelativeFretUpFromIndex(stringNum, note, allFrets.size());
-  }
 }